lobby db, first pass eclipse inspired cleanups
git-svn-id: https://svn.code.sf.net/p/triplea/code/trunk/triplea@1464 a4c711aa-641f-0410-879d-9b464e52b682

+++ src/games/strategy/engine/random/CryptoRandomSource.java
@@ -50,7 +50,7 @@ public class CryptoRandomSource implements IRandomSource
 
     static int byteToIntUnsigned(byte val)
     {
-        return ((int) val) & 0xff;  
+        return val & 0xff;  
     }
     
     public static int[] bytesToInts(byte[] bytes)
